move folder under components
