+++ components/Navbar.vue
@@ -168,7 +168,7 @@ import { get } from 'idb-keyval'
 
 import BasicImage from '@/components/shared/view/BasicImage.vue'
 import ColorModeButton from '@/components/common/ColorModeButton.vue'
-import Identity from '@/components/shared/identity/IdentityIndex.vue'
+import Identity from '@/components/identity/IdentityIndex.vue'
 import NavbarProfileDropdown from '@/components/rmrk/Profile/NavbarProfileDropdown.vue'
 import Search from '@/components/rmrk/Gallery/Search/SearchBar.vue'
