+++ components/landing/CuratedList.vue
@@ -66,7 +66,7 @@ import { CollectionMetadata } from '../rmrk/types'
 
 const components = {
   // Identicon: () => import('@polkadot/vue-identicon'),
-  Identity: () => import('@/components/shared/identity/IdentityIndex.vue'),
+  Identity: () => import('@/components/identity/IdentityIndex.vue'),
 }
 
 const curatedCollection = [
