+++ components/rmrk/Profile/NavbarProfileDropdown.vue
@@ -90,7 +90,7 @@ const components = {
   Avatar,
   ConnectWalletButton: () =>
     import('@/components/shared/ConnectWalletButton.vue'),
-  Identity: () => import('@/components/shared/identity/IdentityIndex.vue'),
+  Identity: () => import('@/components/identity/IdentityIndex.vue'),
   AccountBalance: () => import('@/components/shared/AccountBalance.vue'),
 }
