+++ components/spotlight/SpotlightTable.vue
@@ -216,7 +216,7 @@ import spotlightSoldHistory from '@/queries/rmrk/subsquid/spotlightSoldHistory.g
 type Address = string | GenericAccountId | undefined
 
 const components = {
-  Identity: () => import('@/components/shared/identity/IdentityIndex.vue'),
+  Identity: () => import('@/components/identity/IdentityIndex.vue'),
   Money: () => import('@/components/shared/format/Money.vue'),
   SpotlightDetail: () => import('./SpotlightDetail.vue'),
   Loader: () => import('@/components/shared/Loader.vue'),
